Add pmecc header to bootstrap binary
The bootstrap bin file needs a 52-word header prepended to it when in flash.
In order to support writing using nandwrite and uboot without added
complexity, we'll prepend this header at build time. This also requires
that sam-ba writes flash using the same way we write uboot and kernel; in
other words, we won't use the "Send bootfile" script anymore.

This change uses a static binary file.  It's configurable per board if needed,
and only the wb45n board (at91sam9x5ek) uses it, other configurations would
be unaffected.