Makefile: remove target 'boot' and fixboot script
To make ROM code load at91bootstrap we need put a the file size in
vector table.
As we use .lds to calcuate the _romsize, and put it in the vector
table of generated binary file.
So we don't need 'boot' target and fixboot script to do this any more.

Signed-off-by: Josh Wu <josh.wu@atmel.com>
Acked-by: Nicolas Ferre <nicolas.ferre@atmel.com>