Fix bug in loading images from nandflash
The nandflash driver does not take into account of images that
start from the middle of block when calculating the number of
pages to read. This causes data from the beginning of the first
block to be read due to the read pointer wrapping around after
reaching the end of the block.

Signed-off-by: Dushara Jayasinghe <dushara@successful.com.au>
Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>